+++ data/cube/Cube.js
@@ -6,61 +6,56 @@
  */
 
 
-import {Field} from './Field';
-import {CubeRecord} from './record';
-import {Query} from './Query';
-import {QueryExecutor} from './impl/QueryExecutor';
-import {isString} from 'lodash';
+import {managed} from '@xh/hoist/core';
+import {Query, View, CubeField} from './';
+import {Store} from '../';
 
+import {isEmpty} from 'lodash';
 /**
- * A container for grouping, aggregating, and filtering data on multiple dimensions.
+ * A data store that supports grouping, aggregating, and filtering data on multiple dimensions.
+ *
+ * This object is a wrapper around a Store.   It allows executing queries against that store
+ * that performing filtering, grouping, and aggregating.  It also support the creation of
+ * observable "views" for showing realtime updates to this data..
  */
 export class Cube {
 
-    _fields = null;
-    _records = null;
-    _idSpec = null;
-    _info = null;
-    _lockFn = null;
-
     static RECORD_ID_DELIMITER = '>>';
+
+    @managed
+    store = null;
+    lockFn = null;
+
+    _info = null;
+    _connectedViews = new Set();
 
     /**
      * @param {Object} c - Cube configuration.
-     * @param {(Field[]|Object[])} c.fields - array of Fields / {@see Field} configs.
-     * @param {Object[]} [c.data] - array of raw data.
+     * @param {(CubeField[]|Object[])} - array of CubeFields / {@see CubeField} configs.
+     *      See Store.fields.
+     * @param {Object[]} [c.data] - array of initial raw data.
+     * @param {(function|string)} [c.idSpec] - see Store.idSpec.  Default 'id'
+     * @param {function} [c.processRawData] - see Store.processRawData.
      * @param {Object} [c.info] - map of metadata associated with this data.
-     * @param {(String|function)} [c.idSpec] - property representing unique id of loaded records.
-     * @param {function} [c.lockFn] - function to be called for each node to determine if it should
-     *      be "locked", preventing drilldown into its children. If true returned for a node, no
-     *      drilldown will be allowed, and the row will be marked with a boolean "locked" property.
+     * @param {LockFn} [c.lockFn] - optional function to be called for each node to aggregate to
+     *      determine if it should be "locked", preventing drilldown into its children.
      */
-    constructor({fields, lockFn, idSpec = 'id', data, info}) {
-        this._idSpec = idSpec;
-        this._fields = this.processRawFields(fields);
-        this._records = this.processRawData(data || []);
-        this._info = Object.freeze(info || {});
-        this._lockFn = lockFn;
-    }
-
-    /** @returns {Map} - map of Fields supported by this Cube, by Field name. */
-    get fields() {
-        return this._fields;
-    }
-
-    /** @returns {Field[]} */
-    get fieldList() {
-        return Array.from(this.fields.values());
-    }
-
-    /** @returns {string[]} */
-    get fieldNames() {
-        return Array.from(this.fields.keys());
-    }
-
-    /** @returns {CubeRecord[]} - CubeRecords loaded into this Cube */
-    get records() {
-        return this._records;
+    constructor({
+        fields,
+        idSpec = 'id',
+        processRawData,
+        data = [],
+        info = {},
+        lockFn
+    }) {
+        this.store = new Store({
+            fields: this.parseFields(fields),
+            idSpec,
+            processRawData
+        });
+        this.store.loadData(data);
+        this.lockFn = lockFn;
+        this._info = info;
     }
 
     /** @returns {Object} - optional metadata associated with this Cube at the last data load. */
@@ -68,48 +63,118 @@ export class Cube {
         return this._info;
     }
 
+    //------------------
+    // Querying API
+    //-----------------
+
+    /**
+     * Query the cube.
+     *
+     * This method will return a snapshot of javascript objects representing the filtered
+     * and aggregated data in the query.  In addition to the fields specified in Query, nodes will
+     * each contain a 'cubeLabel' and a 'cubeDimension' property.
+     *
+     * @param {Object} query - Config for query defining the shape of the view.
+     * @returns {Object[]} - data containing the results of the query as a hierarchical set of rows.
+     */
+    executeQuery(query) {
+        query = new Query({...query, cube: this});
+        const view = new View({query}),
+            rows = view.rows;
+
+
+        view.destroy();
+        return rows;
+    }
+
+    /**
+     * Create a View on this data.
+     *
+     * Similar to executeQuery(), but data will be returned as a View which can be
+     * refreshed as the underlying facts in the cube are updated.  Useful for binding
+     * to grids and efficiently displaying changing results in the cube.
+     *
+     * Note: Applications should call the disconnect() or destroy() method on the View
+     * returned when appropriate to avoid unnecessary processing.
+     *
+     * @param {Object} c - config object.
+     * @param {Query} c.query - query to be used to construct this view.
+     * @param {Store} [c.store] - Store to be loaded/reloaded with data from this view.
+     *      To receive data only, use the 'rows' property of the returned object instead.
+     * @param {boolean} [c.connect] - true to update View automatically when data in
+     *      the underlying cube is changed. Default false.
+     * @returns {View}.
+     */
+    createView({query, store, connect = false}) {
+        query = new Query({...query, cube: this});
+        return new View({query, store, connect});
+    }
+
+    //-------------------
+    // Data Loading API
+    //-------------------
     /**
      * Populate this cube with a new dataset.
+     *
+     * This method largely delegates to Store.loadData().  See that method for more
+     * information.
+     *
      * @param {Object[]} rawData - flat array of lowest/leaf level data rows.
      * @param {Object} info - optional metadata to associate with this cube/dataset.
      */
-    async loadDataAsync(rawData, info = {}) {
-        this._records = this.processRawData(rawData);
-        this._info = info;
+    loadData(rawData, info = {}) {
+        this.store.loadData(rawData);
+        this._info = Object.freeze(info);
+        this._connectedViews.forEach(view => view.noteCubeLoaded());
     }
 
     /**
-     * Return grouped and filtered data.
+     * Update this cube with incremental data set changes and/or info.
      *
-     * @param {Object} query - config for a {@see Query}.
-     * @returns {Promise<Object>} - hierarchical representation of filtered and aggregated data, suitable
-     *      for passing directly to a Hoist Store.
+     * This method largely delegates to Store.updateData().  See that method for more
+     * information.
+     *
+     * @param {(Object[]|StoreTransaction)} rawData
+     * @param {Object} info
      */
-    async executeQueryAsync(query) {
-        query = new Query({...query, cube: this});
-        return QueryExecutor.getDataAsync(query);
-    }
+    updateData(rawData, info) {
+        // 1) Process data
+        const changeLog = this.store.updateData(rawData);
 
+        // 2) Process info
+        const infoUpdated = isEmpty(info);
+        if (!isEmpty(info)) {
+            this._info = {...this._info, info};
+        }
+
+        // 3) Notify connected views
+        if (changeLog || infoUpdated) {
+            this._connectedViews.forEach(view => {
+                view.noteCubeUpdated(changeLog, infoUpdated);
+            });
+        }
+    }
 
     //---------------------
     // Implementation
     //---------------------
-    processRawFields(raw) {
-        const ret = new Map();
-        raw.forEach(f => {
-            const field = f instanceof Field ? f : new Field(f);
-            ret.set(field.name, field);
-        });
-        return ret;
+    parseFields(fields = []) {
+        return fields.map(f => f instanceof CubeField ? f : new CubeField(f));
     }
 
-    processRawData(rawData) {
-        return rawData.map(raw => this.createRecord(raw));
-    }
-
-    createRecord(raw) {
-        const {_idSpec} = this;
-        const id = isString(_idSpec) ? raw[_idSpec] : _idSpec(raw);
-        return new CubeRecord(this._fields, raw, id);
+    destroy() {
+        this._connectedViews.forEach(v => v.disconnect());
     }
 }
+
+
+/**
+ * @callback LockFn
+ *
+ * Function to be called for each node to aggregate to determine if it should be "locked",
+ * preventing drilldown into its children. If true returned for a node, no drilldown will be
+ * allowed, and the row will be marked with a boolean "locked" property.
+ *
+ * @param {AggregateRow} row - node to be potentially locked.
+ * @returns boolean
+ */