+++ data/cube/aggregate/UniqueAggregator.js
@@ -10,10 +10,10 @@ import {Aggregator} from '@xh/hoist/data/cube/aggregate/Aggregator';
 
 export class UniqueAggregator extends Aggregator {
 
-    aggregate(records, fieldName) {
-        if (isEmpty(records)) return null;
+    aggregate(rows, fieldName) {
+        if (isEmpty(rows)) return null;
 
-        const val = records[0].get(fieldName);
-        return records.every(it => it.get(fieldName) == val) ? val : null;
+        const val = rows[0][fieldName];
+        return rows.every(it => it[fieldName] === val) ? val : null;
     }
 }